% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_helpers.R
\name{calculate.sonia}
\alias{calculate.sonia}
\title{Run soNNia Selection Analysis}
\usage{
\method{calculate}{sonia}(
  data_folder,
  data_filename,
  pgen_filename,
  organism = "human",
  dataset_type = "TCR",
  n_epochs = 100,
  save_folder = "sonia_output"
)
}
\arguments{
\item{data_folder}{Directory containing data files}

\item{data_filename}{Name of selected TCR data file}

\item{pgen_filename}{Name of background Pgen file}

\item{organism}{Organism: "human" or "mouse"}

\item{dataset_type}{Type of dataset: "TCR" or "BCR"}

\item{n_epochs}{Number of training epochs}

\item{save_folder}{Directory for saving outputs}
}
\value{
List with selection model results
}
\description{
Internal function that calls soNNia via basilisk.
}
\keyword{internal}
