% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.assertSCObject}
\alias{.assertSCObject}
\title{Stop unless the input is a supported single-cell container}
\usage{
.assertSCObject(x, arg = "input")
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Name to use in the error message.}
}
\value{
Invisibly TRUE; called for the error.
}
\description{
Stop unless the input is a supported single-cell container
}
\keyword{internal}
