% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.isSCObject}
\alias{.isSCObject}
\title{Is this a supported single-cell container?}
\usage{
.isSCObject(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
TRUE for SingleCellExperiment or Seurat, FALSE otherwise.
}
\description{
Is this a supported single-cell container?
}
\keyword{internal}
