% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.TCR_summary}
\alias{print.TCR_summary}
\title{Print Method for TCR_summary Objects}
\usage{
\method{print}{TCR_summary}(x, ...)
}
\arguments{
\item{x}{An object of class "TCR_summary"}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Custom print method for TCR repertoire summary objects.
}
