% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huggingModel.R
\name{huggingModel}
\alias{huggingModel}
\title{Initialize a Hugging Face Model and Tokenizer}
\usage{
huggingModel(model_name = "facebook/esm2_t12_35M_UR50D")
}
\arguments{
\item{model_name}{A string specifying the model identifier from the
Hugging Face Hub. For ESM-2 35M, this is "facebook/esm2_t12_35M_UR50D".}
}
\value{
A list containing the R-wrapped Python objects for the 'model'
  and 'tokenizer'.
}
\description{
Downloads or loads a cached pre-trained model and its
  corresponding tokenizer from the Hugging Face Hub.
}
\examples{
\dontrun{
  hf_components <- huggingModel()
  model <- hf_components$model
  tokenizer <- hf_components$tokenizer
}
}
