% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.TCR_summary}
\alias{print.TCR_summary}
\title{Print Method for TCR_summary Objects}
\usage{
\method{print}{TCR_summary}(x, ...)
}
\arguments{
\item{x}{An object of class "TCR_summary"}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Custom print method for TCR repertoire summary objects.
}
\examples{
tcr_data <- data.frame(
  barcode = paste0("cell_", 1:10),
  cdr3_aa = c("CASSLGTGELFF", "CASSIRSSYEQYF", "CASSLGTGELFF",
              "CASSYSTGELFF", "CASSIRSSYEQYF", "CASSLGTGELFF",
              "CASNQGLNEKLFF", "CASSYSTGELFF", "CASSLGTGELFF",
              "CASSIRSSYEQYF"),
  v = rep("TRBV7-2", 10),
  j = rep("TRBJ2-2", 10),
  chain = rep("TRB", 10),
  stringsAsFactors = FALSE
)
summary_obj <- summarizeTCRrepertoire(tcr_data)
print(summary_obj)

}
