% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.TCR_summary}
\alias{print.TCR_summary}
\title{Print method for TCR summary}
\usage{
\method{print}{TCR_summary}(x, ...)
}
\arguments{
\item{x}{TCR_summary object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for TCR summary
}
