% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinEmbeddings.R
\name{proteinEmbeddings}
\alias{proteinEmbeddings}
\title{Get Protein Embeddings from a Model}
\usage{
proteinEmbeddings(
  model,
  tokenized.batch,
  pool = c("none", "mean", "cls"),
  chunk_size = NULL,
  prefer_dtype = c("float16", "bfloat16", "float32"),
  prefer_device = c("auto", "cuda", "mps", "cpu")
)
}
\arguments{
\item{model}{HF model (from AutoModel or similar)}

\item{tokenized.batch}{A *list* of tokenized tensors OR a list of such lists
(i.e., already minibatched). If you pass a single big batch, set chunk_size.}

\item{pool}{One of "mean", "cls", or "none". "mean" is recommended.}

\item{chunk_size}{If tokenized.batch is a single big batch, split it into
chunks of this many sequences. Ignored if you pre-batched upstream.}

\item{prefer_dtype}{One of "float16","bfloat16","float32"}

\item{prefer_device}{One of "auto","cuda","mps","cpu"}
}
\value{
An R matrix [n_sequences x hidden] if pool != "none".
        If pool == "none", returns a list of arrays per chunk.
}
\description{
Applies a pre-trained model to a batch of tokenized sequences
  to generate embeddings.
}
\examples{
  hf_components <- huggingModel()
  sequences <- c("MKTAYIAKQRQISFVKSHFSRQLEERLGLIEVQAPILSRVGDGTQDNLSGAEKAVQVKVKALP",
                 "MKKLFWRAVFLFLLAGLVACSP")
  tokenized_output <- tokenizeSequences(hf_components$tokenizer, sequences)
  model_output <- proteinEmbeddings(hf_components$model, tokenized_output)

}
